refactor: move setting template to new file
